IOSS: Move getline from c to C++ file
